+++ CHANGES.md
@@ -12,8 +12,8 @@ To be released.
 
  -  Added `LastCommit` property to `IBlockMetadata`.  [[#PBFT]]
  -  Added `IBlockPolicy.GetValidators()` method.  [[#PBFT]]
- -  (Libplanet.Net) Removed `SwarmOptions.StaticPeers`.  [[#PBFT]]
  -  Bumped `BlockMetadata.CurrentProtocolVersion` to 4.  [[#PBFT]]
+ -  (Libplanet.Net) Removed `SwarmOptions.StaticPeers`.  [[#PBFT]]
 
 ### Backward-incompatible network protocol changes
 
@@ -27,6 +27,12 @@ To be released.
  -  Added `BlockContent.Propose()` method.  [[#PBFT]]
  -  Added `BlockCommit` class.  [[#PBFT]]
  -  Added `BlockChain.ProposeBlock()` method.  [[#PBFT]]
+ -  Added `Crypto.ConsensusCryptoBackend` class.  [[#PBFT]]
+ -  Added static `ConsensusCryptoBackend` property to `Crypto.CryptoConfig`.
+    [[#PBFT]]
+ -  Added `Crypto.BlsPrivateKey` class.  [[#PBFT]]
+ -  Added `Crypto.BlsPublicKey` class.  [[#PBFT]]
+ -  Added `Crypto.BlsSignature` class.  [[#PBFT]]
  -  (Libplanet.Net) Added `IReactor` interface.  [[#PBFT]]
  -  (Libplanet.Net) Added `ConsensusReactor` class which inherits
     `IReactor` interface.  [[#PBFT]]
@@ -63,6 +69,13 @@ To be released.
 ### Bug fixes
 
 ### Dependencies
+ -  `Libplanet` is now depends on
+    [*Planetarium.Cryptography.BLS12_381* 0.1.10.1][BLS12_381 0.1.10.1].
+    [[#PBFT]]
+
+
+[BLS12_381 0.1.10.1]:
+https://www.nuget.org/packages/Planetarium.Cryptography.BLS12_381
 
 ### CLI tools